requirements.txt
cycler==0.9.0
Django==1.9.7
django-cors-headers==1.1.0
django-rest-swagger==0.3.7
djangorestframework==3.3.3
matplotlib==1.5.1
numpy==1.10.4
oauthlib==1.0.3
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
PyYAML==3.11
requests==2.7.0
requests-oauthlib==0.5.0
scikit-learn==0.17
scipy==0.17.0
six==1.9.0
tabulate==0.7.5
twitter==1.17.1
TwitterAPI==2.3.3.4
virtualenv==14.0.5
wheel==0.24.0
wordcloud==1.1.3
xlwt==1.0.0
